+++ layeredimage/__init__.py
@@ -1,1 +1,2 @@
-""" layeredimage """
+"""Use this module to read, and write to a number of layered image formats
+"""