+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.5 - 2020/04/27
+- Added LSR support
+- Added make.py
+- Bugfix pdn layer offset in group
+- Print more descriptive LayerImage
+- Bugfixes to flattenLayerOrGroup
+
 ## 2020.4.2 - 2020/04/27
 - Update documentation (again) using pydoc-markdown 3